last update, work on the tests

+++ app/controllers/comments_controller.rb
@@ -1,12 +1,16 @@
 class CommentsController < ApplicationController
+  before_action :authenticate_user!
+
   def destroy
     comment = Comment.find(params[:id])
+    authorize comment
     comment.destroy
-    redirect_to send("#{@comment.commentable.class.to_s.downcase}_path", @comment.commentable)
+    redirect_to send("#{comment.commentable.class.to_s.downcase}_path", comment.commentable)
   end
   
   def create
     comment = Comment.new(comment_params)
+    authorize comment
     if comment.save
       flash[:notice] = "Saved"
       redirect_to send("#{comment.commentable.class.to_s.downcase}_path", comment.commentable)
